Add deletion for checkins.
