+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/eventregistration/checkins/CheckIn.kt
@@ -1,5 +1,6 @@
 package de.rki.coronawarnapp.eventregistration.checkins
 
+import de.rki.coronawarnapp.eventregistration.storage.entity.TraceLocationCheckInEntity
 import org.joda.time.Instant
 
 @Suppress("LongParameterList")
@@ -18,4 +19,21 @@ data class CheckIn(
     val checkInEnd: Instant?,
     val targetCheckInEnd: Instant?,
     val createJournalEntry: Boolean
+)
+
+fun CheckIn.toEntity() = TraceLocationCheckInEntity(
+    id = id,
+    guid = guid,
+    version = version,
+    type = type,
+    description = description,
+    address = address,
+    traceLocationStart = traceLocationStart,
+    traceLocationEnd = traceLocationEnd,
+    defaultCheckInLengthInMinutes = defaultCheckInLengthInMinutes,
+    signature = signature,
+    checkInStart = checkInStart,
+    checkInEnd = checkInEnd,
+    targetCheckInEnd = targetCheckInEnd,
+    createJournalEntry = createJournalEntry
 )
