+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/eventregistration/storage/dao/CheckInDao.kt
@@ -3,8 +3,12 @@ package de.rki.coronawarnapp.eventregistration.storage.dao
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
+import androidx.room.Transaction
 import androidx.room.Update
+import de.rki.coronawarnapp.eventregistration.checkins.CheckIn
+import de.rki.coronawarnapp.eventregistration.checkins.toEntity
 import de.rki.coronawarnapp.eventregistration.storage.entity.TraceLocationCheckInEntity
+import de.rki.coronawarnapp.eventregistration.storage.entity.toCheckIn
 import kotlinx.coroutines.flow.Flow
 
 @Dao
@@ -21,6 +25,22 @@ interface CheckInDao {
 
     @Update
     suspend fun update(entity: TraceLocationCheckInEntity)
+
+    @Transaction
+    suspend fun updateEntityById(checkInId: Long, update: (CheckIn?) -> CheckIn?) {
+        val current = entryForId(checkInId)
+
+        val updated = update(current?.toCheckIn()).also {
+            if (it != null && it.id != checkInId) throw UnsupportedOperationException("Can't change entity id: $it")
+        }?.toEntity() ?: return
+
+        if (current != null) {
+            update(updated)
+        } else {
+            if (updated.id == 0L) throw IllegalArgumentException("Adding a new enitity requires default ID 0.")
+            insert(updated)
+        }
+    }
 
     @Query("DELETE FROM checkin")
     suspend fun deleteAll()
